Include relevant type-use annotations in MethodInfo/FieldInfo
First, because that's what the javadoc of MethodInfo advertises.
For example in the javadoc of
org.jboss.jandex.MethodInfo#annotations(org.jboss.jandex.DotName):

     * Retrieves annotations declared on this method, by the name of the
       annotation. This includes annotations which are defined against
method parameters, as
     * well as type annotations declared on any usage within the method
       signature. The <code>target()</code> of the returned annotation
instances may be used to
     * determine the exact location of the respective annotation
       instance.

Notice in particular "as well as type annotations declared on any usage
within the method signature".

The second, perhaps more important reason, is related to index
writing/reading.

In short, putting a reference to the annotation
instance in the method means the annnotation instance will be
written/read together with that method, with the (good) consequence
that org.jboss.jandex.IndexWriterV2#writeTypeTarget will get passed the
enclosing method as the "caller" parameter, and will now correctly set
that method as the enclosing target.

When the annotation instance was not referenced in the method, the
annotation instance was written together with the enclosing type, so
org.jboss.jandex.IndexWriterV2#writeTypeTarget got passed that type as
the "caller" parameter and considered that the type was the enclosing
target, which is wrong.