Add support for strings longer than 64 kB
The `Data{Input,Output}Stream` classes only support [de]serializing
strings up to 64 kB of length, because they store/expect the length
of the string as a `short`. Sometimes, we need to read/write
longer strings. This commit does that by not using the built-in
support for strings, but doing a very similar thing manually.

A string is serialized as a number of bytes of an UTF-8 encoding
of the string (using our custom variable-length encoding of numbers),
followed by the byte array of the UTF-8 encoding of the string.
This only applies in a new persistent index format version. Older
versions are still handled using the old encoding.

Note that while `Data{Input,Output}Stream` uses a custom encoding
for strings, called "modified UTF-8", we use UTF-8 directly. For our
purposes, it shouldn't make a difference.