Fix reading invalid bytecode generated by Kotlin compiler
The JVM specification prohibits methods with more than 255 parameters
(and the number of parameters include the `this` implicit parameter
and even counts `long` and `double` parameters as two!).

Given this limitation, the class file format further only uses 1 byte
to store the number of method parameters on several places.

The Kotlin compiler, however, happily compiles code that includes
methods with more than 255 parameters, and the single-byte values
in the class file are just wrong. Hence, relying on those values
when parsing leads to complete mismatch.

This commit verifies the number of parameters as present in method
signature and when it's higher than 255, it discards the single-byte
value and uses the number of parameters. But all things considered,
this is really a workaround and Kotlin should be fixed.