Propagate type parameter bounds and their annotations
An extra pass over the index is added to `Indexer.complete` that
goes over type parameters of all classes and methods, in topological
order, and for each bound of each type parameter which is itself
a type parameter, copies all the information from that type parameter
declaration (baring the annotations on the bound).

This also includes a partial fixup of unresolved type variables.
If an unresolved type variable occurs in the type parameter bound
position, it is replaced by the corresponding type parameter
(again, baring the annotations on the bound).