make Indexer.index() return void
The `Indexer.index()` method used to return `ClassInfo`, which
assumes that no whole-index post-processing is not necessarsy.
Indeed Jandex doesn't do any such post-processing at the moment,
but it seems pretty obvious that it is necessary. Thus, as a first
step, this commit changes this method to return `void`, so that
all `Indexer` callers are forced to call `complete()` before
being able to access any kind of indexed information.

There are 2 kinds of usages of the `Indexer.index()` result of
the `ClassInfo` type in the Jandex project itself: showing some
summary during batch indexing and creating a single-class index
for tests.

For the first use case, a method `indexWithSummary` is added
that returns `ClassSummary`. This is fine and doesn't affect future
index post-processing, because the summary contains very high-level
information (just a name and number of annotations).

For the second use case, the `IndexingUtil` class is added which
contains utility methods for the most common cases of creating
a single-class index.

Additionally, the `IndexingUtil` class has a `rountrip` method
that serializes the index in memory and deserializes it back.
This is used in several tests, which usually had their own
variation of this code (one even serialized the index to a file).
This commit changes all such places to use this `roundtrip` method.