fix name table serialization
Previously, when serializing the name table, we'd store the "depth"
of the currently serialized name, and during deserialization, we'd
look at the previously deserialized name and take its prefix "depth"
times in hope that we'll obtain the correct prefix. That usually
works, but fails with weird names using the `$` character.

With this commit, we no longer store the "depth" of a name. Instead,
we store an offset to the name table where the prefix is found.
Since names are serialized in order, the prefix must have already
been written. And since the offset is always negative, we actually
store its absolute value to save space (the packed U32 format is
able to store negative values, but it is inefficient).

This change requires persistent format version bump. That has already
happened on the `smallrye` branch, so this commit doesn't bump
again, but note that this commit can't be backported to Jandex 2.x.