add uniform API for annotation access to AnnotationTarget
The `AnnotationTarget` interface gains these methods:

- `boolean hasAnnotation(DotName)`
- `AnnotationInstance annotation(DotName)`
- `Collection<AnnotationInstance> annotations(DotName)`
- `Collection<AnnotationInstance> annotationsWithRepeatable(DotName, IndexView)`
- `Collection<AnnotationInstance> annotations()`
- `boolean hasDeclaredAnnotation(DotName)`
- `AnnotationInstance declaredAnnotation(DotName)`
- `Collection<AnnotationInstance> declaredAnnotationsWithRepeatable(DotName, IndexView)`
- `Collection<AnnotationInstance> declaredAnnotations()`

Many classes that implement `AnnotationTarget` already had a lot
of these methods. First major change here is that now, _all_
implementations of `AnnotationTarget` do. A second major change
is the addition of methods with "declared" in the name: those only
look at annotations present on the annotation target in question
and not on nested annotation targets. Finally, a third major change
is that the collection-returning methods now always return immutable
collections.

The `Type` class is expanded to provide a similar set of methods:

- `boolean hasAnnotation(DotName)`
- `AnnotationInstance annotation(DotName)`
- `Collection<AnnotationInstance> annotationsWithRepeatable(DotName, IndexView)`
- `Collection<AnnotationInstance> annotations()`

The `MethodParameterInfo` class is a first-class citizen now.
It is possible to navigate from `MethodInfo` to all parameters
represented as `MethodParameterInfo`, plus `MethodParameterInfo`
exposes the parameter type. `MethodInfo` also exposes the number
of parameters directly as `parametersCount()`.

All this happens only on the API level, internal representation
of anything doesn't change. There are two breaking changes:

- the `ClassInfo.annotations()` method no longer returns a `Map`,
  it returns a `List` now; the original method is still available
  as `annotationsMap()`
- the `MethodInfo.parameters()` method no longer returns a `List<Type>`,
  it returns a `List<MethodParameterInfo>` now; the original method
  is still available as `parameterTypes()`

Additionally, all the `ClassInfo.classAnnotation*` methods are
deprecated in favor of `declaredAnnotation*`.