add ClassInfo.memberClasses
This allows navigating from `ClassInfo` to its member classes.
The `memberClasses()` method returns a set of `DotName` objects,
because member classes don't necessarily have to be present in
the index. It is a responsibility of the caller to retrieve
a `ClassInfo` of each member class from the index and handle
possible absence.

This change requires persistent format version bump. That has already
happened on the `smallrye` branch, so this commit doesn't bump
again, but note that this commit can't be backported to Jandex 2.x.