uniform handling of implicitly declared (aka mandated) and synthetic parameters
The most common parameter-related methods on `MethodInfo` now pretend
that mandated and synthetic parameters don't exist. Parameter names
and annotations are indexed in bytecode using the same assumption, so
this change also allows small cleanup of type annotation handling.

This is a semi-breaking change, because Jandex has already been doing
it for methods that have the `Signature` attribute in bytecode. Note
that javac emits the `Signature` attribute not just for methods that
use a type variable in their signature, but also for constructors of
enums or local/anonymous classes. In case a method doesn't have the
`Signature` attribute, Jandex will employ a simple heuristic that
covers the most important cases: constructors of enums and non-static
member classes. Constructors of local/anonymous classes compiled with
ecj are not covered perfectly, but that seems to be acceptable.

There are extra methods to access the full list of parameter types as
present in the method descriptor, but those types can't be correlated
with parameter names or annotations.

This change requires persistent format version bump. That has already
happened on the `smallrye` branch, so this commit doesn't bump again,
but note that this commit can't be backported to Jandex 2.x.