Fix wrong type parameter sharing in generic signature parser
The generic signature parser used to forget previously encountered
element-level type parameters when parsing a signature of a new
element, and previously encountered class-level type parameters when
parsing a signature of a new class. This leaves a hole: previously
encountered element-level type parameters could be used when
parsing a signature of a new class. When parsing a signature of
a new class, previously encountered element-level type parameters
must also be forgotten, to prevent wrong sharing.