Workaround for wrong type annotations on compact record constructor
It seems javac has a bug when emitting a compact constructor of a record
that declares a component whose type has type annotations. For example:

    public record MyRecord(List<@NotEmpty String> list) {
        public MyRecord {
        }
    }

In this case, javac emits a type annotation whose `target_type` is 0x13
(that is, the type annotation is supposed to target a field). That is wrong.

My guess is that javac simply copies the type from the implicitly declared
component field, together with type annotations, into the implicitly declared
parameter list of the compact constructor. (On the other hand, type annotations
on the implicitly declared accessor methods are correct.)

With this commit, Jandex ignores such erroneous type annotations.