Propagate type annotations on all type variables + improve Type.toString()
This commit builds on existing support for propagating type annotations
on type parameter bounds. After the existing type annotations on type
parameter bounds propagation pass, a second pass is added that propagates
type annotations on all remaining type variables.

As part of this commit, the string representation of types is improved.
The `Type.toString()` method now has an overload that takes a "simple"
flag. This flag is used to make type variables not append their bounds
into their string representation. With this, the string representation
of declarations and composite types look much closer to the corresponding
Java source code.