Add proper support for recursive type parameters
This commit adds a new kind of type, a `TypeVariableReference`, which is used
in a class/method list of type parameters to represent a type variable that
isn't fully defined yet. For simple recursive type parameters, such as
`T extends Comparable<T>`, this naturally means that type variables occurring
in their own definitions are represented as references. For mutually recursive
type parameters, type variables that occur in their definitions or before their
definitions are represented as references. A type variable reference can be
`follow()`-ed to obtain the underlying type variable.

The `TypeVariableReference` is mutable, initially points to `null` and must
later be patched. This occurs on several occasions:

1. When indexing a class, generic signatures are parsed, which includes
   resolution of type variables (where unresolved type variables that are
   found recursive are replaced by references) and then patching of newly
   introduced references.
2. Whe completing an index, type annotations are propagated, which includes
   creating deep copies of type variables that include references. To create
   such deep copies, new references are introduced, to avoid sharing. Such
   new references are later patched.
3. When deserializing an index, types are read in topological order (if a type
   is composite, it is read after its constituent types), which also requires
   patching references (because they are read earlier).

This change requires persistent format version bump. That has already
happened on the `smallrye` branch, so this commit doesn't bump again,
but note that this commit can't be backported to Jandex 2.x.