Add String- and Class-accepting shortcuts
In Jandex 3.0.0, we added shortcut methods to `IndexView` so that users
don't have to always create an instance of `DotName`. These methods
accept `String` (which is wrapped into `DotName`) and even `Class<?>`
(whose `getName()` is wrapped into `DotName`).

In this commit, we add similar shortcuts to `DotName.createSimple()`
and all annotation access methods on `AnnotationTarget`.