Improve how the `test-data` module is built using JDK 17
Previously, the `test-data` module was always built using a second JDK
downloaded during build using a Maven plugin. This is hostile to downstream
rebuilds in an air-gapped environment and makes the release process
more complex and fragile for no good reason.

With this commit, a second JDK is only downloaded if the "current" JDK
is older than 17. If the "current" JDK is 17+, it is used to build all
modules, including `test-data`. We still want to make sure that the core
Jandex library, as well as the Maven plugin, produce Java 8 bytecode.
For that reason, this commit adds an automated test to both modules which
verifies that the class file version of some class is `52` (Java 8).

Also, this commit changes the JDK version used for release to 17, to
eat our own dog food.