Use the `--release` option when compiling on JDK 9+
When compiling with JDK 8, we still need to use `-source` and `-target`,
because JDK 8 doesn't know the `--release` option. However, with JDK 9+,
it is better to use `--release`, which not only instructs the compiler
to treat the source code as Java 8 and emit Java 8 bytecode, but also
compiles against JDK 8 core library signatures.