Ignore invalid generic signatures during indexing
In case an invalid generic signature appears in the bytecode, indexing
fails with an exception. This is unfriendly, because ECJ emits invalid
generic signatures on pretty basic code (a synthetic static method
generated for a lambda), so ECJ-generated code cannot be indexed.

This commit simply ignores a generic signature that cannot be parsed.