Improve representation of array types
The Jandex representation of array types is different from the Java
language representation; it is compressed into a constituent type
and a number of dimensions. The constituent type may be another
array type in case type annotations are present. Often, they are not,
in which case the constituent type corresponds to the element type;
but when they are, there's no direct correspondence to the Java
language representation of array types.

Up until now, the constituent type of an array type has even been
called its "component", which is incorrect. This commit renames that
to "constituent" and adds the `constituent()` method. The existing
`component()` method is obviously retained for backwards compatibility,
but becomes deprecated.

Additionally, the `elementType()`, `componentType()` and
`deepDimensions()` methods are added that directly correspond
to the concepts from the Java language representation.