Simplify implementation of JandexReflection
The `loadRawType` method used to inspect the type and have a special
branch for each kind of types. This is useless, because for vast majority
of types, the `Type.name()` method returns a name that can be directly
used with `Class.forName()`. This includes array types, which means
that `loadRawType` no longer needs to allocate an array to obtain
its class object. The only types for which extra handling is still
required are: the void pseudo-type, primitive types, and unresolved
type variables.