fix null handling in the interning pool
This commit undoes many changes made incorrectly in a previous commit
where the underlying `Object[]` in the interning pool was changed to
`E[]` (where `E` is the interning pool's type parameter). This was
incorrect because the table may contain a special `NULL` marker object
that is never of type `E`.

Further, to fix `null` handling, the equality and hash code functions
are improved to treat the `NULL` marker object specially. The goal is
that the `NULL` marker object never makes it to the type-specific
subclasses, because that would lead to a class cast error.

The problem with interning `null` hasn't been noticed, because interning
the `null` value is in fact very rare. Still, it is valid and must work.