allow primitive type signatures on fields
The grammar for generic signatures in the JVMS is pretty clear that
a signature for a field is a reference type signature. However,
there are class files in the wild that contain a primitive type
signature for a field. This commit makes that acceptable in Jandex,
even though it technically isn't valid bytecode.