add StackedIndex
A stacked index is a composition of multiple indices with overlay semantics.
Overlaying is done on class granularity. When the composition doesn't contain
duplicate classes (multiple classes with the same name), a `StackedIndex`
behaves just like a `CompositeIndex`, but when duplicates appear, the behavior
of `StackedIndex` is actually well defined.

This commit also adds a test for index navigation methods from `IndexView`,
as those have not been tested at all. Some tests reveal inconsistencies
in how `CompositeIndex` behaves in presence of duplicates, which is not nice,
but this commit intentionally doesn't modify `CompositeIndex` behavior.