Fix class ordering when propagating type annotations, take two
To propagate type annotations across nested classes, we sort the entire
set of indexed classes so that outer classes come before their inner
classes. We don't really care about ordering between classes that are
not members of the same nest (using the JEP 181 terminology). That is,
we only need a partial order on classes.

However, to perform that sort, we use the sorting routine in the JDK, which
uses the order defined by a `Comparator`, and that order must be total.

To define a total order, this commit takes a completely opposite approach.
We compute a nesting level for each class (where a top-level class has
a nesting level of 0, a class nested in a top-level class has a nesting
level of 1, etc.) and sort by the nesting level. For equal nesting levels,
we sort by the class name. This means that all top-level classes come first,
all classes nested in top-level classes come second, etc.

To verify that a given `Comparator` establishes a total order for a given
set of values, this commit includes a simple utility class `TotalOrderChecker`,
which was used to reproduce (and understand) the problem of previous
implementation of this ordering.