improve Index[View].getKnownUsers()
Previously, the `getKnownUsers()` method only considered a class
as used in another class when it occured in the list of class
references in the other class's constant pool.

With this commit, a class is also considered as used in another class
when it occurs in the other class's signature (superclass type,
superinterface types, type parameters, permitted subclasses),
in the signatures of the other class's methods (return type,
parameter types, exception types, type parameters), or in the types
of the other class's fields and record components.