Avoid having a ton of empty HashMaps in memory
I'm working on a heap dump and Eclipse MAT complained that there was
a lot of empty HashMaps around. I had a closer look and a lot of them
were due to the Map kept in ClassInfo and this one looked like the most
plausible culprit. Now, in Quarkus, the Jandex index is just around
for building the application, but it looks like an easy win.