fix losing type annotations with different visibility
In case a single type annotation target (such as a `MethodInfo`)
has multiple type annotations, some of them runtime-retained and
some class-retained, all annotations of one kind are lost.

This is because `Indexer.processTypeAnnotations()` was written with
the assumption that it is only called once for any given target.
This was true in Jandex 2, which only supported runtime-retained
annotations, but since Jandex 3.0, class-retained annotations are
supported as well. Hence, the method can legally be called twice
for any given target.

Fortunately, this issue doesn't exist for non-type annotations.