add ClassInfo.enclosingClassAlways()
The `enclosingClassAlways()` returns an enclosing class not only for
member classes, but also for local and anonymous classes, regardless
of their declaration place. They may be declared in the body of a method
or constructor, or in a static, instance or field initializer.

The behavior of `enclosingClass()` doesn't change.

Note that this commit increments the persistent format version.