fix ClassInfo.simpleName()
The `simpleName()` method used to return a wrong result in case of top-level
classes with dollar `$` in their names. For example, for the following class

```java
public class Top$Level {
}
```

the `ClassInfo.simpleName()` method used to return `Level`. This commit fixes
that by only calling `name.local()` in case the `name` is componentized (which
it always is, so just to be sure) and is _not_ marked inner. If it is marked
inner, the `simpleName()` method calls `name.withoutPackagePrefix()`.

It would be possible to call `withoutPackagePrefix()` always, but that method
allocates in case of componentized names, which we want to avoid if possible.