fix comment in ClassInfo.simpleName()
Commit message of one of the previous commits, as well as comment
in `ClassInfo.simpleName()`, say that `simpleName()` could just always
use `name.withoutPackagePrefix()`. That is not true, this would lead
to wrong simple names in case of nested classes. It would only work
for top-level classes. This commit fixes the comment.