intern EquivalenceKeys for primitive types and class types of java.*
Primitive types and `java.*` class types are very commonly used and there
is a limited number of them. Therefore, we can intern them safely (growth
of the interning cache is naturally bounded) and improve allocation rate
significantly.

Further, this commit removes all calls to `Objects.hash()`, because each call
allocates an array. It is especially useless when called with just one element.

Co-authored-by: Guillaume Smet <guillaume.smet@gmail.com>