Avoid using componentized DotNames for usual suspects
We are building StringBuilders again and again in DotName#toString()
while I think it has little value to use componentized DotNames for
these very common DotNames as long as we reuse the same instances
everywhere, which is the case, if I'm not mistaken.

To be honest, I wonder if in some cases, it wouldn't have value to allow
prepopulating the NameTable with additional common entries before the
Index is read/built. Typically for the ArC annotations. I'm not sure
it's either feasible or desirable but I thought I would put this idea
there.