Make index serializations reproducible
Jandex indexes produced with the current implementation are not reproducible, in the sense that multiple serializations of the same index produce the same binary representation. Having reproducible binary representations is important for reproducible builds, in order to verify a build result by rerunning a build.

The only reason that the serialized representation is not reproducible is due to letting `Set`s and `Map` dictate the order in which elements like classes, modules and usages are serialized.

The changes to make index serialization reproducible are luckily not very intrusive. It is mainly implemented via a new `IndexWriteView` class, which extracts the necessary collections from an `Index` and provides sorted `List`s used during serialization. Another part is the processing order of the "class elements" member-classes, permitted-subclasses and annotation-entries, which are now "guarded" with a sort operation.

No index format changes are necessary, all indexes serialized with `IndexWriterV2` benefit from this change. `IndexWriterV1` (for the not even anymore documented index format versions) has not been changed.

Fixes #568