skip annotations when computing a hash code of fields, methods and record components
Computing a hash code of an `AnnotationInstance` is relatively expensive,
but makes little sense as part of a hash code of `FieldInternal`, `MethodInternal`
and `RecordComponentInternal`. This is because there should never be fields,
methods and record components that are the same except annotations. Further,
there's plenty other information in the field/method/record component hash code,
so the impact on hash collisions should be minimal.